Removed GAMESCOPE_NV12_COLORSPACE default
This was being set in UpdateCompatEnvVars(), which is called after the environment variable is already read. The default in rendervulkan.cpp if the environment variable isn't set is BT.709 full-range, which is a sane default, so this isn't necessary.

(cherry picked from commit 751b97ccf75b5ffdb58aeb423eeee578405b426e)