layer: Force up swapchain minImageCount
The Thing: Remastered (2958970) always creates a swapchain with 2 images in vkCreateSwapchain's minImageCount and does not respect the surface capabilities minImageCount.

This causes the game to run at half framerate when SteamOS is not compositing, because we only return the image back to the game when the new one is presented on screen successfully which causes a huge stall in GPU work for the game.

Forcing the image count to respect the surface capabilities at creation time fixes this problem and allows the game to run at a solid 90fps.

(cherry picked from commit 91f975562958f33a78f77f996f098e1518b6c088)