steamcompmgr: track FSR state with preemptive upscaling
g_bFSRActive was only being applied to the first frame of preemptive
upscaling, which meant that the FSR badge would quickly flicker from on
to off when preemptive upscaling was active.

To account for this, track the state of preemptive upscaling and use
it to activate g_bFSRActive when applicable.

(cherry picked from commit 81e40911e425c41071f3f684eba76b154b25f7af)